Improve host (or1k cpu) accesses to SDRAM
This changes the qmem_bridge from being a 32-16 bit bridge to a
32-32 bridge.
To achieve this, a couple of changes are made to the sdram
controller.
1) a simple 'burst' cache is added to the host 'port' of the
controller. All it does is save the 3 extra 16-bit words that
are fetched on all accesses into a buffer.
2) The writes are changed from single accesses to burst writes.
That way the full 32-bit host writes can be written in a single
sdram cycle. All other writes in the three extra slots are
performed as no-ops, by holding dqm high on them.

The slave side clock of the qmem bridge has also been changed
from clk_7 to clk_114, that way early acks (cache hits and writes)
can propagate propagate faster out on the 50MHz clock side of
the bridge.